refactor BlogItem
